+++ backend/user-center/src/main/java/vn/winter/usercenter/user/UserController.java
@@ -4,13 +4,10 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import vn.winter.usercenter.otp.OTP;
+import vn.winter.usercenter.user.dto.ActiveUserDto;
 import vn.winter.usercenter.user.dto.UserSignInDto;
 import vn.winter.usercenter.user.dto.UserSignUpDto;
-import vn.winter.usercenter.util.OtpType;
-import vn.winter.usercenter.util.Util;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController()
@@ -38,7 +35,11 @@ public class UserController {
         return this.userService.signIn(userSignInDto);
     }
 
-    // Active user
+    @PatchMapping("/active-user")
+    public ResponseEntity<Object> activeUser(@Valid @RequestBody ActiveUserDto activeUserDto) {
+        return this.userService.activeUser(activeUserDto);
+    }
+
     // Re-send OTP
     // Forget password
     // Change password