+++ backend/user-center/src/main/java/vn/winter/usercenter/user/dto/ActiveUserDto.java
@@ -1,0 +1,21 @@
+package vn.winter.usercenter.user.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ActiveUserDto {
+    @NotBlank
+    @Email
+    private String email;
+
+    @NotBlank
+    private String otp;
+}