+++ backend/user-center/src/main/java/vn/winter/usercenter/user/dto/UserDto.java
@@ -1,9 +1,6 @@
 package vn.winter.usercenter.user.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.io.Serializable;
 import java.time.LocalDate;
@@ -14,6 +11,7 @@ import java.util.Date;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class UserDto implements Serializable {
     private long userId;
     private String fullName;