+++ backend/user-center/src/main/java/vn/winter/usercenter/util/ResponseMap.java
@@ -15,6 +15,18 @@ public class ResponseMap {
         return this;
     }
 
+    public ResponseMap setMessage(String message) {
+        return setValue("message", message);
+    }
+
+    public ResponseMap setStatus(int status) {
+        return setValue("status", status);
+    }
+
+    public ResponseMap setData(Object data) {
+        return setValue("data", data);
+    }
+
     public Map<String, Object> build() {
         // Clone the map
         Map<String, Object> cloneMap = new HashMap<>(this.responseMap);